Handle builds without commits gracefully

+++ lib/integrity/build.rb
@@ -22,7 +22,9 @@ module Integrity
     has 1,     :commit
 
     before :destroy do
-      commit.destroy!
+      if commit
+        commit.destroy!
+      end
     end
 
     def run
@@ -66,10 +68,18 @@ module Integrity
     end
 
     def sha1
-      commit.identifier
+      if commit
+        commit.identifier
+      else
+        '(commit is missing)'
+      end
     end
 
     def sha1_short
+      unless commit
+        return '(commit is missing)'
+      end
+
       unless sha1
         return "This commit"
       end
@@ -78,15 +88,29 @@ module Integrity
     end
 
     def message
-      commit.message || "message not loaded"
+      if commit
+        commit.message || "message not loaded"
+      else
+        '(commit is missing)'
+      end
     end
 
     def author
-      (commit.author || Author.unknown).name
+      if commit
+        (commit.author || Author.unknown).name
+      else
+        '(commit is missing)'
+      end
     end
 
     def committed_at
-      commit.committed_at
+      if commit
+        commit.committed_at
+      else
+        # Returning a string here might break something, if it does
+        # change to something like beginning of epoch or nil
+        '(commit is missing)'
+      end
     end
 
     def status
